% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_HCDR3s.R
\name{extract_HCDR3s}
\alias{extract_HCDR3s}
\title{Extract and clean rabbit HCDR3s}
\usage{
extract_HCDR3s(aa)
}
\arguments{
\item{aa}{AA string set of translated sequences}
}
\description{
For a subset of cell lines
}
