% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_embedded_primer.R
\name{filter_embedded_primer}
\alias{filter_embedded_primer}
\title{Filter embedded CH1 primer from reads}
\usage{
filter_embedded_primer(DNA, primer, read_length)
}
\arguments{
\item{DNA}{sequence}

\item{primer}{sequence of reverse primer}

\item{read_length}{read length of single end read}
}
\description{
Identify reads containing correct reverse primer, filter them, then trim reads
such that HCDR3 is in correct reading frame
}
