% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primer.R
\name{get_primer}
\alias{get_primer}
\title{Get reverse primer sequences for subset of HCDR3s}
\usage{
get_primer(sequence, annealing = 62)
}
\arguments{
\item{sequence}{sequence to trim from 3' end to obtain primer. reverse complement of HCDR3 template}

\item{annealing}{annealing temperature of primers, default = 62}
}
\description{
Trims reverse complemented HCDR3 DNA sequence to appropriate primer length, based upon annealing temperature calculation
}
\examples{
get_primer("CCAGATGTCGGCATTATTAGGCCATCTTCTGGCACAGAAATAGGTGGCCGTGTCCGAGGCTGTCAGACTGTTCATTTTCAGATCCATCGTGGTCGAGGTTTTGGA", annealing = 62)

### "CCAGATGTCGGCATTATTAGGCCATCTTCT"
}
