% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_rabbit_HCDR3s.R
\name{splice_rabbit_HCDR3s}
\alias{splice_rabbit_HCDR3s}
\title{Splice out rabbit HCDR3 sequence}
\usage{
splice_rabbit_HCDR3s(
  aa,
  upstream_junction = ".TYFC|A.YFC|AT.FC|ATY.C",
  downstream_junction = "WG.G"
)
}
\arguments{
\item{aa}{AA string set of translated sequences}

\item{upstream_junction}{protein sequence REGEX upstream of rabbit HCDR3}

\item{downstream_junction}{protein sequence REGEX downstream of rabbit HCDR3}
}
\description{
Splices using amino acid regular expressions that recognize rabbit HCDR3 junctional sequences
}
