% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_calc.R
\name{tm_calc}
\alias{tm_calc}
\title{Calculate tm of DNA oligo sequence}
\usage{
tm_calc(sequence)
}
\arguments{
\item{sequence}{primer sequence, as string}
}
\description{
Calculate tm of DNA oligo sequence
}
\examples{
tm_calc("GCCCTTGGTGGAGGC")

#### 46.15714
}
