% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_HCDR3.R
\name{differential_HCDR3}
\alias{differential_HCDR3}
\title{Calculate depmap score mean for depmap cell lines, and N per gene}
\usage{
differential_HCDR3(
  experimental_fastq_files,
  control_fastq_files,
  primer = "GCCCTTGGTGGAGGC"
)
}
\arguments{
\item{experimental_fastq_files}{vector of names of experimental fastq files for analysis. Ordered to be paired with control_fastq_files}

\item{control_fastq_files}{vector of names of control fastq files for analysis. Ordered to be paired with experimental_fastq_files}

\item{primer}{CH1 hybridizing reverse primer.}
}
\description{
Function to conduct differential HCDR3 abundance analysis of single end MiSeq reads
from FBCseq experiment
}
