% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_embedded_primer.R
\name{filter_embedded_primer}
\alias{filter_embedded_primer}
\title{Calculate depmap score mean for depmap cell lines, and N per gene}
\usage{
filter_embedded_primer(DNA, primer)
}
\arguments{
\item{primer}{sequence of reverse primer}

\item{seq}{sequence}

\item{read_length}{Miseq read length - default = 250}
}
\description{
For a subset of cell lines
}
